# Speeding up sum reductions in ADADELTA by using Tensor Cores