--- a/Makefile.Cuda
+++ b/Makefile.Cuda
@@ -56,6 +56,10 @@ endif
 
 ifeq ($(TENSOR), ON)
 	NVTENSOR=-DUSE_NVTENSOR
+
+	ifeq ($(TCEC), ON)
+		NVTENSOR+=-I./wmma_extension/include -DUSE_TCEC
+	endif
 endif
 
 BIN := $(wildcard $(TARGET)*)