Hi,

This PR aims to increase the performance of the CUDA version by leveraging the Tensor Cores Units (TCU) present in recent NVIDIA GPUs.

The idea is to re-implement the sum reductions as matrix operations (i.e., by using [NVIDIA Warp Matrix Functions](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#warp-matrix-functions)), which can be offloaded to TCUs.

Experiments on A100 GPU (`make DEVICE=GPU TESTLS=ad NUMWI=64 TARGETS=80 test`):

| Docking time |  Original|  Tensor|
|:-:|-:|-:|
| In seconds| 0.8 | 0.6 |

Experiments on RTX3050Ti GPU (`make DEVICE=GPU TESTLS=ad NUMWI=64 TARGETS=86 test`):

| Docking time |  Original|  Tensor|
|:-:|-:|-:|
| In seconds| 2.4 | 1.7 |

The baseline implementation for this PR has been taken from this paper: [_Accelerating Drug Discovery in AutoDock-GPU with Tensor Cores_](https://link.springer.com/chapter/10.1007/978-3-031-39698-4_41). The contribution of both authors, Gabin Schieffer (@gabin-s) and Ivy Peng (@bopkth), is acknowledged in this PR as well:

```
Schieffer, Gabin, and Peng, Ivy. "Accelerating Drug Discovery in AutoDock-GPU with Tensor Cores."
In European Conference on Parallel Processing, pp. 608-622. Cham: Springer Nature Switzerland, 2023.
```