@L30nardoSV 2024-01-11
> @L30nardoSV Thank you very much, I am currently testing. Please encapsulate the code a bit and make it a compile option so older Cuda versions and cards still compile and run.

OK, let me know if the `TENSOR` directive in commit 10b07fa6a suffices