@atillack 2024-01-11
@L30nardoSV I tested on one of our Nvidia Quadro RTX A5000 cards and I do see a nice speedup for the 3ce3 example input:

Docking time of the PR with `make DEVICE=GPU TESTLS=ad NUMWI=64 TARGETS=86 test` is 0.70 seconds vs 0.90 seconds (this does use the heuristics and autostop by default).

To evaluate a bit further I used Diogo's test set of 42 ligands, here are the results:

Reference:
| Path | NUMWI | AutoStop & Heuristics |  overall evals | energy | rmsd | docking | idle |
|:----:|:--------:|:-------:|:------:|:------:|:-----:|:--------:|:----:|
| OpenCL | 128 | no, 2.5M evals | 105382018 | 36  /  42 good | 36  /  42 good | 91.69 s | 0.17 s |
| Cuda | 128 | no, 2.5M evals | 105331182 | 36  /  42 good | 36  /  42 good  | 90.75 s | 0.32 s |
| Cuda | 64 | no, 2.5M evals | 105404961  | 36  /  42 good | 35  /  42 good | 106.24 s | 7.93 s |
| OpenCL | 128 | yes | 84026192 | 37  /  42 good | 37  /  42 good | 187.82 s |0.21 s |
| Cuda | 128 | yes | 80037847 | 38  /  42 good | 38  /  42 good | 184.85 s | 0.20 s |
| Cuda | 64 | yes | 84684628 | 36  /  42 good | 38  /  42 good | 233.39 s | 8.13 s |

This PR:
| Path | NUMWI | AutoStop & Heuristics |  overall evals | energy | rmsd | docking | idle |
|:----:|:--------:|:-------:|:------:|:------:|:-----:|:--------:|:----:|
| OpenCL | 128 | no, 2.5M evals | 105362595 | 38  /  42 good | 36  /  42 good | 92.33 s | 0.21 s |
| Cuda | 128 | no, 2.5M evals | 105177642 | 35  /  42 good | 36  /  42 good | 100.71 s | 0.20 s |
| Cuda | 64 | no, 2.5M evals | 105197433 | 35  /  42 good | 38  /  42 good |112.48 s | 0.19 s |
| OpenCL | 128 | yes | 86495325 | 37  /  42 good | 38  /  42 good | 192.71 s | 0.21 s |
| Cuda | 128 | yes | 71419809 | 33  /  42 good | 37  /  42 good | 182.30 s | 0.21 s |
| Cuda | 64 | yes | 65754981 | 34  /  42 good | 37  /  42 good | 214.60 s | 0.22 s |

For multiple differently sized ligands with the typical settings It turns out for larger systems the speedup can turn into a slowdown.

It looks like the average number of evals w/ AutoStop changed in the PR which could potentially point to a minute difference in calculation (i did test multiple times to make sure this wasn't just an unlucky run).

@diogomart Please run your E50 tests for the Cuda version.