@diogomart 2024-01-12
Unfortunately, algorithmic performance is worse.

![79f13c7-ocl-128wi_vs_PR252-10b07fa-cuda-tensor-128wi-overlap](https://github.com/ccsb-scripps/AutoDock-GPU/assets/1564797/99dc5847-2755-4e70-932d-b2db88288d5a)