@L30nardoSV 2024-02-16
@atillack 

Can you please check commit [b2ab3fe](https://github.com/ccsb-scripps/AutoDock-GPU/pull/252/commits/b2ab3fe6b79e08b7c1af5f11233070ac6c42d025) that incorporates an WMMA Extension for single precision matmul on Tensor Cores + error correction (TCEC)?

`make DEVICE=GPU TESTLS=ad NUMWI=64 TARGETS=80 TENSOR=ON TCEC=ON test`

Ref: https://github.com/wmmae/wmma_extension/blob/main/docs/mma_f32.md