@atillack 2024-02-20
@L30nardoSV I ran the newest version and here are the results (with OpenCL from before as comparison, note: i compiled w/o OVERLAP so the last column takes a bit longer, but compute times are unaffected):
| Path | NUMWI | AutoStop & Heuristics |  overall evals | energy | rmsd | docking | idle |
|:----:|:--------:|:-------:|:------:|:------:|:-----:|:--------:|:----:|
| OpenCL | 128 | yes | 86495325 | 37  /  42 good | 38  /  42 good | 192.71 s | 0.21 s |
| Cuda | 128 | yes | 88164353| 36  /  42 good | 38  /  42 good | 194.13 s | 7.74 s |
| Cuda | 64 | yes | 77884078 | 37  /  42 good | 37  /  42 good | 214.27 s | 21.54 s |