@diogomart 2024-07-16
@L30nardoSV sorry for the long delay. I don't see an improvement, unfortunately. Very similar results to the previous commit.
![79f13c7-ocl-128wi_vs_PR252-b2ab3fe-cuda-tensor-128wi-overlap](https://github.com/user-attachments/assets/33cf82eb-1921-47ba-b65e-e7670299f108)