@L30nardoSV 2024-07-17
@diogomart thanks! 
I just want to make sure: did you compile with both `TENSOR=ON TCEC=ON` ?