@diogomart 2024-07-17
With TENSOR=ON yes, but not with TCEC=ON