@L30nardoSV 2024-07-17
> With TENSOR=ON yes, but not with TCEC=ON

Can you please try again with `TCEC=ON`?