@diogomart 2024-07-17
TCEC fixed it :+1: 

![79f13c7-ocl-128wi_vs_PR252-b2ab3fe-cuda-tensor-tcec-128wi-overlap](https://github.com/user-attachments/assets/c870622b-1694-4436-8fb1-d940e3556f5f)