@diogomart 2024-07-17
I think there's a small improvement over plain CUDA but OpenCL is still the fastest.

![PR255-7007db8-cuda-128wi-overlap--PR252-b2ab3fe-cuda-tensor-tcec-128wi-overlap runtime](https://github.com/user-attachments/assets/9743fb03-97e0-4bc2-85cd-223ae74b1486)
![PR255-7007db8-cuda-128wi-overlap--PR255-7007db8-ocl-128wi-overlap runtime](https://github.com/user-attachments/assets/c9e5fb91-e8b6-4971-84c3-1f801de45cb8)