@atillack 2024-07-17
What is the overall runtime divided by the overall number of evals? (and what is the std.error?)