@atillack 2024-07-17
(and by runtime, i really mean docking time - although with overlap, runtime is probably good enough)