@atillack 2024-07-17
Also, did you run on the same type of GPU?