@diogomart 2024-07-17
no, this was a mix of rtx5000 and rtx6000