@atillack 2024-07-17
here is the data for the reduced set of 42:

| Path | NUMWI | AutoStop & Heuristics |  overall evals | energy | rmsd | docking | idle |
|:----:|:--------:|:-------:|:------:|:------:|:-----:|:--------:|:----:|
| Cuda | 128 | yes | 80037847 | 38  /  42 good | 38  /  42 good | 184.85 s | 0.20 s |
| Cuda | 128 | yes | 88164353| 36  /  42 good | 38  /  42 good | 194.13 s | 7.74 s |

In other words, i've seen 184,850,000 microseconds / 80,037,847 = 2.31 microseconds per eval vs 2.20 microseconds per eval for this PR which would be about 5% faster ... except that that's right around what i would estimate AutoStop to fluctuate by ...

I'll run with a fixed number of evals later to get one more data point and try to get on one of our newest cards as well to see if newer tensor cores might be beneficial.