@L30nardoSV 2024-07-17
Thanks @diogomart and @atillack 

Glad to see that by enabling the error correction code (via TCEC) the docking quality is back to normal.

> I'll run with a fixed number of evals later to get one more data point and try to get on one of our newest cards as well to see if newer tensor cores might be beneficial.

I look forward to seeing those numbers :)