@diogomart 2024-07-17
Each marker is system and aggregates data from 32000 to 8M evals.
![time_per_eval_tensor_vs_opencl](https://github.com/user-attachments/assets/75cd6242-6335-488e-b749-bde11a4f7612)